+++ app/shared/app-data/src/commonMain/kotlin/data/models/subject/SubjectProgressInfo.kt
@@ -11,7 +11,6 @@ package me.him188.ani.app.data.models.subject
 
 import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.Stable
-import me.him188.ani.app.data.models.SubjectInfo
 import me.him188.ani.app.data.models.episode.EpisodeInfo
 import me.him188.ani.app.data.models.episode.isKnownCompleted
 import me.him188.ani.app.data.repository.EpisodeCollectionInfo
