+++ app/shared/src/commonMain/kotlin/ui/subject/collection/SubjectCollectionsColumn.kt
@@ -58,9 +58,10 @@ import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.itemContentType
 import androidx.paging.compose.itemKey
 import androidx.window.core.layout.WindowWidthSizeClass
-import me.him188.ani.app.data.models.SubjectCollectionInfo
 import me.him188.ani.app.data.models.episode.EpisodeInfo
 import me.him188.ani.app.data.models.subject.SubjectAiringInfo
+import me.him188.ani.app.data.models.subject.SubjectCollectionInfo
+import me.him188.ani.app.data.models.subject.SubjectInfo
 import me.him188.ani.app.data.models.subject.SubjectProgressInfo
 import me.him188.ani.app.data.repository.EpisodeCollectionInfo
 import me.him188.ani.app.tools.paging.exceptions
@@ -388,7 +389,7 @@ private fun testSubjectCollection(
     episodes: List<EpisodeCollectionInfo>,
     collectionType: UnifiedCollectionType,
 ): SubjectCollectionInfo {
-    val subjectInfo = me.him188.ani.app.data.models.SubjectInfo.Empty.copy(
+    val subjectInfo = SubjectInfo.Empty.copy(
         id,
         nameCn = "中文条目名称",
         name = "Subject Name",
